refactor: extract Session's buy/sell/launch family into session.purchases
