+++ FlipcashTests/SessionTests.swift
@@ -471,90 +471,6 @@ struct SessionTests {
 // MARK: -
 
 @MainActor
-@Suite("Session.buy verified state")
-struct SessionBuyVerifiedStateTests {
-
-    private static let staleAmount = ExchangedFiat(
-        onChainAmount: TokenAmount(quarks: 1_000_000, mint: .usdf),
-        nativeAmount: .usd(1),
-        currencyRate: .oneToOne
-    )
-
-    @Test("buy throws verifiedStateStale when the provided state is past clientMaxAge")
-    func buy_throwsStale() async {
-        let session = Session.unverifiedMock
-        let stale = VerifiedState.stale(bonded: false)
-
-        do {
-            _ = try await session.buy(
-                amount: Self.staleAmount,
-                verifiedState: stale,
-                of: .usdf
-            )
-            Issue.record("Expected verifiedStateStale to be thrown")
-        } catch Session.Error.verifiedStateStale {
-            // expected
-        } catch {
-            Issue.record("Unexpected error thrown: \(error)")
-        }
-    }
-
-    @Test("buyNewCurrency throws verifiedStateStale when the provided state is past clientMaxAge")
-    func buyNewCurrency_throwsStale() async {
-        let session = Session.unverifiedMock
-        let stale = VerifiedState.stale(bonded: false)
-
-        do {
-            _ = try await session.buyNewCurrency(
-                amount: Self.staleAmount,
-                feeAmount: Self.staleAmount,
-                verifiedState: stale,
-                mint: .usdf
-            )
-            Issue.record("Expected verifiedStateStale to be thrown")
-        } catch Session.Error.verifiedStateStale {
-            // expected
-        } catch {
-            Issue.record("Unexpected error thrown: \(error)")
-        }
-    }
-}
-
-// MARK: -
-
-@MainActor
-@Suite("Session.sell verified state")
-struct SessionSellVerifiedStateTests {
-
-    private static let amount = ExchangedFiat(
-        onChainAmount: TokenAmount(quarks: 1_000_000, mint: .usdf),
-        nativeAmount: .usd(1),
-        currencyRate: .oneToOne
-    )
-
-    @Test("sell throws verifiedStateStale when the provided state is past clientMaxAge")
-    func sell_throwsStale() async {
-        let session = Session.unverifiedMock
-        let stale = VerifiedState.stale(bonded: false)
-
-        do {
-            _ = try await session.sell(
-                amount: Self.amount,
-                verifiedState: stale,
-                in: .usdf
-            )
-            Issue.record("Expected verifiedStateStale to be thrown")
-        } catch Session.Error.verifiedStateStale {
-            // expected
-        } catch {
-            Issue.record("Unexpected error thrown: \(error)")
-        }
-    }
-}
-
-// MARK: -
-
-@MainActor
 @Suite("Session.balances(for:) USDF inclusion")
 struct SessionBalancesUSDFInclusionTests {
